Remove old low level functions
