+++ include/mgard-x/CompressionLowLevel/CMakeLists.txt
@@ -1,7 +1,6 @@
 list(APPEND MGARD_X_HEADER
-    ${CMAKE_CURRENT_SOURCE_DIR}/CompressionLowLevel.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/CompressionLowLevel.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/CompressionLowLevelWorkspace.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/Compressor.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/Compressor.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/NormCalculator.hpp
     )
 set(MGARD_X_HEADER ${MGARD_X_HEADER} PARENT_SCOPE)
