+++ include/mgard-x/Lossless/ParallelHuffman/GetCodebook.hpp
@@ -5,6 +5,7 @@
  * Date: March 17, 2022
  */
 
+#include "HuffmanWorkspace.hpp"
 #include "FillArraySequence.hpp"
 #include "GenerateCL.hpp"
 #include "GenerateCW.hpp"
