+++ include/mgard-x/Quantization/LinearQuantization.hpp
@@ -680,211 +680,6 @@ private:
   SubArray<1, QUANTIZED_INT, DeviceType> outliers;
 };
 
-template <DIM D, typename T, typename DeviceType>
-void LinearQuanziation(
-    Hierarchy<D, T, DeviceType> &hierarchy, Array<D, T, DeviceType> &in_array,
-    Config config, enum error_bound_type type, T tol, T s, T norm,
-    CompressionLowLevelWorkspace<D, T, DeviceType> &workspace, int queue_idx) {
-  SubArray in_subarray(in_array);
-  bool prep_huffman =
-      config.lossless != lossless_type::CPU_Lossless; // always do Huffman
-  SIZE total_elems = hierarchy.total_num_elems();
-  SubArray<2, SIZE, DeviceType> level_ranges_subarray(hierarchy.level_ranges());
-  SubArray<3, T, DeviceType> level_volumes_subarray(
-      hierarchy.level_volumes(false));
-
-  T *quantizers = new T[hierarchy.l_target() + 1];
-  calc_quantizers<D, T>(total_elems, quantizers, type, tol, s, norm,
-                        hierarchy.l_target(), config.decomposition, true);
-  MemoryManager<DeviceType>::Copy1D(workspace.quantizers_subarray.data(),
-                                    quantizers, hierarchy.l_target() + 1,
-                                    queue_idx);
-  LENGTH zero = 0;
-  MemoryManager<DeviceType>::Copy1D(
-      workspace.huffman_workspace.outlier_count_subarray.data(), &zero, 1,
-      queue_idx);
-
-  SubArray<1, QUANTIZED_INT, DeviceType> *quantized_linearized_v_host = nullptr;
-  SubArray<1, QUANTIZED_INT, DeviceType> *quantized_linearized_v = nullptr;
-
-  if (config.reorder) { // only if we need linerization
-    quantized_linearized_v_host =
-        new SubArray<1, QUANTIZED_INT, DeviceType>[hierarchy.l_target() + 1];
-    SIZE *ranges_h = level_ranges_subarray.dataHost();
-    SIZE last_level_size = 0;
-    for (SIZE l = 0; l < hierarchy.l_target() + 1; l++) {
-      SIZE level_size = 1;
-      for (DIM d = 0; d < D; d++) {
-        level_size *= ranges_h[(l + 1) * D + d];
-      }
-      quantized_linearized_v_host[l] = SubArray<1, QUANTIZED_INT, DeviceType>(
-          {level_size - last_level_size},
-          workspace.quantized_subarray(last_level_size));
-      last_level_size = level_size;
-    }
-
-    MemoryManager<DeviceType>::Malloc1D(quantized_linearized_v,
-                                        hierarchy.l_target() + 1, queue_idx);
-    DeviceRuntime<DeviceType>::SyncDevice();
-    MemoryManager<DeviceType>::Copy1D(quantized_linearized_v,
-                                      quantized_linearized_v_host,
-                                      hierarchy.l_target() + 1, queue_idx);
-    DeviceRuntime<DeviceType>::SyncDevice();
-  }
-
-  Timer timer;
-  if (log::level & log::TIME)
-    timer.start();
-
-  bool done_quantization = false;
-  while (!done_quantization) {
-    DeviceLauncher<DeviceType>::Execute(
-        LevelwiseLinearQuantizerKernel<D, T, MGARDX_QUANTIZE, DeviceType>(
-            level_ranges_subarray, hierarchy.l_target(),
-            workspace.quantizers_subarray, level_volumes_subarray, s,
-            config.huff_dict_size, in_subarray, workspace.quantized_subarray,
-            prep_huffman, config.reorder, quantized_linearized_v,
-            workspace.huffman_workspace.outlier_count_subarray,
-            workspace.huffman_workspace.outlier_idx_subarray,
-            workspace.huffman_workspace.outlier_subarray),
-        queue_idx);
-
-    MemoryManager<DeviceType>::Copy1D(
-        &workspace.huffman_workspace.outlier_count,
-        workspace.huffman_workspace.outlier_count_subarray.data(), 1,
-        queue_idx);
-    DeviceRuntime<DeviceType>::SyncQueue(queue_idx);
-    if (workspace.huffman_workspace.outlier_count <=
-        workspace.huffman_workspace.outlier_subarray.shape(0)) {
-      // outlier buffer has sufficient size
-      done_quantization = true;
-      if (log::level & log::TIME) {
-        timer.end();
-        timer.print("Quantization");
-        timer.clear();
-      }
-      log::info("Outlier ratio: " +
-                std::to_string(workspace.huffman_workspace.outlier_count) +
-                "/" + std::to_string(total_elems) + " (" +
-                std::to_string((double)100 *
-                               workspace.huffman_workspace.outlier_count /
-                               total_elems) +
-                "%)");
-    } else {
-      log::info("Not enough workspace for outliers. Re-allocating to " +
-                std::to_string(workspace.huffman_workspace.outlier_count));
-      workspace.huffman_workspace.outlier_idx_array =
-          Array<1, LENGTH, DeviceType>(
-              {(SIZE)workspace.huffman_workspace.outlier_count});
-      workspace.huffman_workspace.outlier_array =
-          Array<1, QUANTIZED_INT, DeviceType>(
-              {(SIZE)workspace.huffman_workspace.outlier_count});
-      workspace.huffman_workspace.outlier_idx_subarray =
-          SubArray(workspace.huffman_workspace.outlier_idx_array);
-      workspace.huffman_workspace.outlier_subarray =
-          SubArray(workspace.huffman_workspace.outlier_array);
-      workspace.huffman_workspace.outlier_count_array.memset(0);
-    }
-  }
-  if (config.reorder) {
-    delete[] quantized_linearized_v_host;
-    MemoryManager<DeviceType>::Free(quantized_linearized_v);
-  }
-
-  delete[] quantizers;
-}
-
-template <DIM D, typename T, typename DeviceType>
-void LinearDequanziation(
-    Hierarchy<D, T, DeviceType> &hierarchy, Array<D, T, DeviceType> &out_array,
-    Config config, enum error_bound_type type, T tol, T s, T norm,
-    CompressionLowLevelWorkspace<D, T, DeviceType> &workspace, int queue_idx) {
-  SIZE total_elems = hierarchy.total_num_elems();
-  out_array.resize(hierarchy.level_shape(hierarchy.l_target()));
-  SubArray<D, T, DeviceType> out_subarray(out_array);
-  MemoryManager<DeviceType>::Copy1D(
-      workspace.huffman_workspace.outlier_count_subarray.data(),
-      &workspace.huffman_workspace.outlier_count, 1, queue_idx);
-  SubArray<2, SIZE, DeviceType> level_ranges_subarray(hierarchy.level_ranges());
-  SubArray<3, T, DeviceType> level_volumes_subarray(
-      hierarchy.level_volumes(true));
-
-  bool prep_huffman = config.lossless != lossless_type::CPU_Lossless;
-
-  T *quantizers = new T[hierarchy.l_target() + 1];
-  calc_quantizers<D, T>(total_elems, quantizers, type, tol, s, norm,
-                        hierarchy.l_target(), config.decomposition, false);
-  MemoryManager<DeviceType>::Copy1D(workspace.quantizers_subarray.data(),
-                                    quantizers, hierarchy.l_target() + 1,
-                                    queue_idx);
-  DeviceRuntime<DeviceType>::SyncQueue(queue_idx);
-
-  SubArray<1, QUANTIZED_INT, DeviceType> *quantized_linearized_v_host = nullptr;
-  SubArray<1, QUANTIZED_INT, DeviceType> *quantized_linearized_v = nullptr;
-  if (config.reorder) { // only if we need linerization
-    quantized_linearized_v_host =
-        new SubArray<1, QUANTIZED_INT, DeviceType>[hierarchy.l_target() + 1];
-    SIZE *ranges_h = level_ranges_subarray.dataHost();
-    SIZE last_level_size = 0;
-    for (SIZE l = 0; l < hierarchy.l_target() + 1; l++) {
-      SIZE level_size = 1;
-      for (DIM d = 0; d < D; d++) {
-        level_size *= ranges_h[(l + 1) * D + d];
-      }
-      quantized_linearized_v_host[l] = SubArray<1, QUANTIZED_INT, DeviceType>(
-          {level_size - last_level_size},
-          workspace.quantized_subarray(last_level_size));
-      last_level_size = level_size;
-    }
-
-    MemoryManager<DeviceType>::Malloc1D(quantized_linearized_v,
-                                        hierarchy.l_target() + 1, queue_idx);
-    DeviceRuntime<DeviceType>::SyncDevice();
-    MemoryManager<DeviceType>::Copy1D(quantized_linearized_v,
-                                      quantized_linearized_v_host,
-                                      hierarchy.l_target() + 1, queue_idx);
-    DeviceRuntime<DeviceType>::SyncDevice();
-  }
-
-  Timer timer;
-  if (log::level & log::TIME)
-    timer.start();
-
-  if (prep_huffman && workspace.huffman_workspace.outlier_count) {
-    DeviceLauncher<DeviceType>::Execute(
-        OutlierRestoreKernel<D, T, DeviceType>(
-            workspace.quantized_subarray,
-            workspace.huffman_workspace.outlier_count,
-            workspace.huffman_workspace.outlier_idx_subarray,
-            workspace.huffman_workspace.outlier_subarray),
-        queue_idx);
-  }
-
-  DeviceLauncher<DeviceType>::Execute(
-      LevelwiseLinearQuantizerKernel<D, T, MGARDX_DEQUANTIZE, DeviceType>(
-          level_ranges_subarray, hierarchy.l_target(),
-          workspace.quantizers_subarray, level_volumes_subarray, s,
-          config.huff_dict_size, out_array, workspace.quantized_subarray,
-          prep_huffman, config.reorder, quantized_linearized_v,
-          workspace.huffman_workspace.outlier_count_subarray,
-          workspace.huffman_workspace.outlier_idx_subarray,
-          workspace.huffman_workspace.outlier_subarray),
-      queue_idx);
-
-  DeviceRuntime<DeviceType>::SyncQueue(queue_idx);
-  if (log::level & log::TIME) {
-    timer.end();
-    timer.print("Dequantization");
-    timer.clear();
-  }
-
-  if (config.reorder) {
-    delete[] quantized_linearized_v_host;
-    MemoryManager<DeviceType>::Free(quantized_linearized_v);
-  }
-  delete[] quantizers;
-}
-
 template <DIM D, typename T, typename Q, typename DeviceType>
 class LinearQuantizer : public QuantizationInterface<D, T, Q, DeviceType> {
 public:
