+++ src/mgard-x/CompressionLowLevel/CMakeLists.txt
@@ -1,5 +1,4 @@
-MgardXGenerateSourceAllCombinations("Compress")
-MgardXGenerateSourceAllCombinations("Decompress")
+MgardXGenerateSourceAllCombinations("Compressor")
 set(MGARD_X_SERIAL_SRC ${MGARD_X_SERIAL_SRC} PARENT_SCOPE)
 set(MGARD_X_OPENMP_SRC ${MGARD_X_OPENMP_SRC} PARENT_SCOPE)
 set(MGARD_X_CUDA_SRC ${MGARD_X_CUDA_SRC} PARENT_SCOPE)
