--- src/mgard-x/CompressionLowLevel/Compress.cpp.in
@@ -1,24 +1,0 @@
-/*
- * Copyright 2022, Oak Ridge National Laboratory.
- * MGARD-X: MultiGrid Adaptive Reduction of Data Portable across GPUs and CPUs
- * Author: Jieyang Chen (chenj3@ornl.gov)
- * Date: March 17, 2022
- */
-
-#include "mgard-x/CompressionLowLevel/CompressionLowLevel.hpp"
-// clang-format off
-namespace mgard_x {
-
-template void
-compress<@NUM_DIM@, @DATA_TYPE@, @DEVICE_TYPE@>(
-    Hierarchy<@NUM_DIM@, @DATA_TYPE@, @DEVICE_TYPE@> &hierarchy,
-    Array<@NUM_DIM@, @DATA_TYPE@, @DEVICE_TYPE@> &original_array,
-    enum error_bound_type type, @DATA_TYPE@ tol, @DATA_TYPE@ s,
-    @DATA_TYPE@ &norm, Config config, 
-    CompressionLowLevelWorkspace<@NUM_DIM@, @DATA_TYPE@, @DEVICE_TYPE@>& workspace,
-    Array<1, Byte, @DEVICE_TYPE@> &compressed_array);
-
-template class Compressor<@NUM_DIM@, @DATA_TYPE@, @DEVICE_TYPE@>;
-
-} // namespace mgard_x
-// clang-format on
